+++ Changelog.md
@@ -5,3 +5,5 @@ Changelog
 * :feature:`1` add did retrieve api using private key.
 * :feature:`2` add api to sign message .
 * :feature:`2` add api to verify message if signed by public address .
+* :feature:`3` add Did setting info api .
+* :feature:`3` add Did getting info from transaction .